+++ classes/output/core/course_renderer.php
@@ -279,17 +279,15 @@ class course_renderer extends \core_course_renderer {
         }
 
         if ($type == 2) {
+            $content .= html_writer::start_tag('div', array('class' => 'coursebox-content'));
             $content .= html_writer::start_tag('a', array(
-                'class' => 'coursebox-content',
                 'href' => new moodle_url('/course/view.php', array('id' => $course->id))
             ));
             $coursename = $chelper->get_course_formatted_name($course);
             $content .= html_writer::tag('h3', $coursename, array('class' => $course->visible ? '' : 'dimmed'));
+            $content .= html_writer::end_tag('a');
         }
         $content .= html_writer::start_tag('div', array('class' => 'summary'));
-        if (ISSET($coursename)) {
-            $content .= html_writer::tag('p', html_writer::tag('strong', $coursename));
-        }
         // Display course summary.
         if ($course->has_summary()) {
             $summs = $chelper->get_course_formatted_summary($course, array('overflowdiv' => false, 'noclean' => true,
@@ -329,7 +327,7 @@ class course_renderer extends \core_course_renderer {
             }
         }
         if ($type == 2) {
-            $content .= html_writer::end_tag('a');
+            $content .= html_writer::end_tag('div');
             // End coursebox-content.
         }
         $content .= html_writer::tag('div', '', array('class' => 'boxfooter')); // Coursecat.